linux-x86_64: change unsigned long to unsigned int - long is 64 bits on x86_64 where sha1 wants 32 bits - should not affect 32 bit builds

+++ sha1.c
@@ -23,12 +23,12 @@ A million repetitions of "a"
 
 /* Hash a single 512-bit block. This is the core of the algorithm. */
 
-void SHA1Transform(unsigned long state[5], unsigned char buffer[64])
+void SHA1Transform(unsigned int state[5], unsigned char buffer[64])
 {
-    unsigned long a, b, c, d, e;
+    unsigned int a, b, c, d, e;
     typedef union {
         unsigned char c[64];
-        unsigned long l[16];
+        unsigned int l[16];
     } CHAR64LONG16;
     CHAR64LONG16* block;
 #ifdef SHA1HANDSOFF
@@ -116,7 +116,7 @@ unsigned int i, j;
 
 void SHA1Final(unsigned char digest[DIGEST_SIZE], SHA1_CTX* context)
 {
-unsigned long i, j;
+unsigned int i, j;
 unsigned char finalcount[8];
 
     for (i = 0; i < 8; i++) {
